Add files via upload
These are data files for the clustering code. This can be used to reproduce all the clustering results from the paper. OFCCaMKII_populationdata is the trial averaged fluorescence of CaMKIIa expressing neurons in vmOFC during Pavlovian conditioning. OFCVTA_populationdata is the trial averaged fluorescence of vmOFC neurons projecting to VTA during Pavlovian conditioning. The other files are intermediate processing files during the analysis in the clustering code